Update K-Means Clustering.ipynb
Running through seeds 150-250 (+1) took 3 hours and 8 minutes, but we have confirmed that we've found the best two seeds possible. I'm locked in.

Having some second guesses about whether or not I've implemented k-means as efficiently as I could have, but if it took 3.25 hours to get through 302 it shouldn't take much longer than 4 hours to create the k means model for 350. Which at this point sounds pretty reasonable considering I was considering a 24 hour runtime.

After creating the model I'll be working on creating visualizations and testing the model with the test data