Update K-Means Clustering.ipynb
The last thing I wanted to do was compare the model's performance  on the training data to its performance on our reserved testing data, to see how well the model was successful in generalizing.

After creating averages of the losses to allow them to be comared, I wasn't satisfied just eyeballing the loss values when I was comparing between the test data and the model data. So I went ahead and found an easy way to implement a statistical ttest with scipy, and ran the test finding that there was not a statistically significant difference between our pivot k 225, or the full list of average losses for each k.

And that's it for this file. all set. ggs