Update Loading Optimized K-Means Model.ipynb
I started doing some graph plotting stuff here, now that it's confirmed that the loading worked and I wanted to make sure we had all of the variables we needed. But I really want to organize all the drawings into a beautified results document. so instead of doing it here I'm going to move on to one last colab file.