Update RESULTS - Plotting K-Means Model Results.ipynb
Got the plot set up for interactions but found that plotly and ipywidgets were having a weird interaction where the plot wasn't displaying.

Found a stackoverflow answer that says that the new versions are causing issues, and added an install to do plotly 5.10

Looking good now! got the initial angle set up but want to see if I can make it so it doesn't reset every time it changes a setting.